+++ src/lips.js
@@ -466,7 +466,9 @@
     Pair.prototype.toString = function() {
         var arr = ['('];
         if (this.car !== undefined) {
-            if (typeof this.car === 'string') {
+            if (typeof this.car === 'function') {
+                arr.push('<#function ' + (this.car.name || 'anonymous') + '>');
+            } else if (typeof this.car === 'string') {
                 arr.push(JSON.stringify(this.car));
             } else if (typeof this.car !== 'undefined') {
                 arr.push(this.car);
@@ -523,12 +525,25 @@
     };
 
     // ----------------------------------------------------------------------
-    // :: Environment constructor (parent argument is optional)
+    // :: Environment constructor (parent and name arguments are optional)
     // ----------------------------------------------------------------------
-    function Environment(obj, parent) {
+    function Environment(obj, parent, name) {
         this.env = obj;
         this.parent = parent;
+        this.name = name;
     }
+    // ----------------------------------------------------------------------
+    Environment.prototype.inherit = function(obj, name) {
+        if (typeof obj === 'string') {
+            name = obj;
+            obj = {};
+        }
+        if (!name) {
+            name = 'child of ' + (this.name || 'unknown');
+        }
+        return new Environment(obj || {}, this, name);
+    };
+    // ----------------------------------------------------------------------
     Environment.prototype.get = function(symbol) {
         if (symbol instanceof Symbol) {
             if (typeof this.env[symbol.name] !== 'undefined') {
@@ -559,6 +574,7 @@
             }
         }
     };
+    // ----------------------------------------------------------------------
     Environment.prototype.set = function(name, value) {
         this.env[name] = value;
     };
@@ -584,7 +600,7 @@
     function let_macro(asterisk) {
         return new Macro(function(code, dynamic_scope) {
             var args = this.get('list->array')(code.car);
-            var env = new Environment({}, this);
+            var env = this.inherit('let');
             return new Promise((resolve) => {
                 var promises = [];
                 var i = 0;
@@ -603,7 +619,7 @@
                         resolve(new Quote(evaluate(
                             output,
                             env,
-                            dynamic_scope ? env : undefined
+                            dynamic_scope ? env : dynamic_scope
                         )));
                     } else {
                         var value = evaluate(
@@ -784,9 +800,9 @@
         // ------------------------------------------------------------------
         'let': let_macro(false),
         // ------------------------------------------------------------------
-        'begin': new Macro(function(code) {
+        'begin': new Macro(function(code, dynamic_scope) {
             var arr = this.get('list->array')(code);
-            return arr.reduce((_, code) => evaluate(code, this), 0);
+            return arr.reduce((_, code) => evaluate(code, this, dynamic_scope), 0);
         }),
         // ------------------------------------------------------------------
         timer: new Macro(function(code) {
@@ -823,7 +839,9 @@
             }
             if (code.car instanceof Symbol) {
                 if (value instanceof Promise) {
-                    value.then(value => this.env[code.car.name] = value);
+                    return value.then(value => {
+                        this.env[code.car.name] = value;
+                    });
                 } else {
                     this.env[code.car.name] = value;
                 }
@@ -850,7 +868,7 @@
         lambda: new Macro(function(code, dynamic_scope) {
             var self = this;
             return function(...args) {
-                var env = new Environment({}, dynamic_scope ? this : self);
+                var env = (dynamic_scope ? this : self).inherit('lambda');
                 var name = code.car;
                 var i = 0;
                 var value;
@@ -878,14 +896,14 @@
                         name = name.cdr;
                     }
                 }
-                return evaluate(code.cdr.car, env);
+                return evaluate(code.cdr.car, env, env);
             };
         }),
         // ------------------------------------------------------------------
         defmacro: new Macro(function(macro) {
             if (macro.car.car instanceof Symbol) {
                 this.env[macro.car.car.name] = new Macro(function(code) {
-                    var env = new Environment({}, this);
+                    var env = new Environment({}, this, 'defmacro');
                     var name = macro.car.cdr;
                     var arg = code;
                     while (true) {
@@ -1352,7 +1370,7 @@
             this.set(code.car, value);
             return value;
         })
-    });
+    }, undefined, 'global');
 
     // ----------------------------------------------------------------------
     // source: https://stackoverflow.com/a/4331218/387194
@@ -1409,7 +1427,22 @@
 
     // ----------------------------------------------------------------------
     function evaluate(code, env, dynamic_scope) {
-        env = env || global_env;
+        /*
+        if (code instanceof Pair) {
+            if (code.car.name) {
+                console.log(code.car.name);
+            }
+            console.log({
+                env: env ? env.name : undefined,
+                dynamic: dynamic_scope ? dynamic_scope.name : undefined,
+                code: code && code.toString()
+            });
+        }*/
+        if (env === true) {
+            env = dynamic_scope = global_env;
+        } else {
+            env = env || global_env;
+        }
         var value;
         if (typeof code === 'undefined') {
             return;
@@ -1480,8 +1513,37 @@
     }
 
     // ----------------------------------------------------------------------
-    function exec(string, env) {
-        return parse(tokenize(string)).map((code) => evaluate(code, env));
+    function exec(string, env, dynamic_scope) {
+        if (env === true) {
+            env = dynamic_scope = global_env;
+        } else {
+            env = env || global_env;
+        }
+        var list = parse(tokenize(string));
+        return new Promise((resolve, reject) => {
+            var results = [];
+            (function recur() {
+                function next(value) {
+                    results.push(value);
+                    recur();
+                }
+                var code = list.shift();
+                if (!code) {
+                    resolve(results);
+                } else {
+                    try {
+                        var result = evaluate(code, env, dynamic_scope);
+                    } catch (e) {
+                        return reject(e);
+                    }
+                    if (result instanceof Promise) {
+                        result.then(next).catch(reject);
+                    } else {
+                        next(result);
+                    }
+                }
+            })();
+        });
     }
 
     // ----------------------------------------------------------------------